Move windows version things in its own file
